Add configuration for tests
This introduces an extensible configuration object as part of the abstract test class.
This enables us to customize test runs by adjusting properties in the configuration.
The class itself is extensible if the need for more customizations should arise.

It also cleans up some of the fields in the abstract test class.